+++ src/baseclasses.py
@@ -1,5 +1,6 @@
-import os
-import xml.etree.ElementTree as ET
+from src.utils import GameListHandler
+from src.utils import FileHandler
+from src.utils import StringHandler
 
 
 class Game:
@@ -7,17 +8,17 @@ class Game:
 
     def __init__(self, filename=None):
         self.filename = filename
-        self.path = None
-        self.name = None
-        self.sortname = None
-        self.description = None
 
+        self.path = None
         self.boxart = None
         self.image = None
         self.marquee = None
         self.thumbnail = None
         self.video = None
 
+        self.name = None
+        self.sortname = None
+        self.description = None
         self.releasedate = None
         self.developer = None
         self.publisher = None
@@ -30,204 +31,175 @@ class Game:
         self.lastplayed = None
         self.md5 = None
 
+    def str_append(self, game_var, text):
+        ush = StringHandler()
+        s = ''
+        if game_var is not None:
+            s += '\n    ' + ush.str_green(text)
+            s += str(game_var)
+        return s
+
     def __str__(self):
-        u = Utils()
-        str = u.str_blue('Filename: ') + self.filename
+        ush = StringHandler()
+        s = ush.str_blue('Filename: ')
+        s += str(self.filename)
 
-        if self.path is not None:
-            str += '\n\t' + u.str_green('Path: ') + self.path
+        s += self.str_append(self.path, 'Path: ')
+        s += self.str_append(self.name, 'Name: ')
+        s += self.str_append(self.desc, 'Description: ')
+        s += self.str_append(self.sortname, 'Sortname: ')
+        s += self.str_append(self.boxart, 'Boxart: ')
+        s += self.str_append(self.image, 'Image: ')
+        s += self.str_append(self.marquee, 'Marquee: ')
+        s += self.str_append(self.thumbnail, 'Thumbnail: ')
+        s += self.str_append(self.video, 'Video: ')
+        s += self.str_append(self.releasedate, 'ReleaseDate: ')
+        s += self.str_append(self.developer, 'Developer: ')
+        s += self.str_append(self.publisher, 'Publisher: ')
+        s += self.str_append(self.players, 'Players: ')
+        s += self.str_append(self.core, 'Core: ')
+        s += self.str_append(self.rating, 'Rating: ')
+        s += self.str_append(self.playcount, 'Playcount: ')
+        s += self.str_append(self.lastplayed, 'LastPlayed: ')
+        s += self.str_append(self.md5, 'MD5: ')
+        s += self.str_append(self.genre, 'Genre: ')
 
-        if self.name is not None:
-            str += '\n\t' + u.str_green('Name: ') + self.name
+        return s + '\n'
 
-        if self.sortname is not None:
-            str += '\n\t' + u.str_green('Sortname: ') + self.sortname
+    def load_media_from_gamelist(self, gamelist, systempath):
+        xmlg = gamelist.get_game_by_filename(self.filename)
+        if gamelist.game_tag_text_exist(xmlg, 'path'):
+            if gamelist.filename_equals_gamepath(self.filename,
+                                                 xmlg.find('path').text):
 
-        if self.boxart is not None:
-            str += '\n\t' + u.str_green('Boxart: ') + self.boxart
+                fh = FileHandler()
 
-        if self.image is not None:
-            str += '\n\t' + u.str_green('Image: ') + self.image
+                if gamelist.game_tag_text_exist(xmlg, 'path'):
+                    path = fh.path([systempath, xmlg.find('path').text])
+                    if fh.check_file(path):
+                        self.path = xmlg.find('path').text
 
-        if self.marquee is not None:
-            str += '\n\t' + u.str_green('Marquee: ') + self.marquee
+                if gamelist.game_tag_text_exist(xmlg, 'boxart'):
+                    path = fh.path([systempath, xmlg.find('boxart').text])
+                    if fh.check_file(path):
+                        self.boxart = xmlg.find('boxart').text
 
-        if self.thumbnail is not None:
-            str += '\n\t' + u.str_green('Thumbnail: ') + self.thumbnail
+                if gamelist.game_tag_text_exist(xmlg, 'image'):
+                    path = fh.path([systempath, xmlg.find('image').text])
+                    if fh.check_file(path):
+                        self.image = xmlg.find('image').text
 
-        if self.video is not None:
-            str += '\n\t' + u.str_green('Video: ') + self.video
+                if gamelist.game_tag_text_exist(xmlg, 'marquee'):
+                    path = fh.path([systempath, xmlg.find('marquee').text])
+                    if fh.check_file(path):
+                        self.marquee = xmlg.find('marquee').text
 
-        return str + '\n'
+                if gamelist.game_tag_text_exist(xmlg, 'thumbnail'):
+                    path = fh.path([systempath, xmlg.find('thumbnail').text])
+                    if fh.check_file(path):
+                        self.thumbnail = xmlg.find('thumbnail').text
 
-    def import_info(self, gamelist):
-        game = gamelist.get_game_by_filename(self.filename)
-        if game:
-            self.path = './'+self.filename
+                if gamelist.game_tag_text_exist(xmlg, 'video'):
+                    path = fh.path([systempath, xmlg.find('video').text])
+                    if fh.check_file(path):
+                        self.video = xmlg.find('video').text
 
-            if game.find('name') is not None:
-                self.name = game.find('name').text
+    def load_info_from_gamelist(self, gamelist):
+        xmlg = gamelist.get_game_by_filename(self.filename)
+        if xmlg:
+            if gamelist.game_tag_text_exist(xmlg, 'name'):
+                self.name = xmlg.find('name').text
 
-            if game.find('sortname') is not None:
-                self.sortname = game.find('sortname').text
+            if gamelist.game_tag_text_exist(xmlg, 'desc'):
+                self.desc = xmlg.find('desc').text
 
-            if game.find('desc') is not None:
-                self.description = game.find('desc').text
+            if gamelist.game_tag_text_exist(xmlg, 'sortname'):
+                self.sortname = xmlg.find('sortname').text
 
-            if game.find('boxart') is not None:
-                self.boxart = game.find('boxart').text
+            if gamelist.game_tag_text_exist(xmlg, 'description'):
+                self.description = xmlg.find('description').text
 
-            if game.find('image') is not None:
-                self.image = game.find('image').text
+            if gamelist.game_tag_text_exist(xmlg, 'releasedate'):
+                self.releasedate = xmlg.find('releasedate').text
 
-            if game.find('marquee') is not None:
-                self.marquee = game.find('marquee').text
+            if gamelist.game_tag_text_exist(xmlg, 'developer'):
+                self.developer = xmlg.find('developer').text
 
-            if game.find('thumbnail') is not None:
-                self.thumbnail = game.find('thumbnail').text
+            if gamelist.game_tag_text_exist(xmlg, 'publisher'):
+                self.publisher = xmlg.find('publisher').text
 
-            if game.find('video') is not None:
-                self.video = game.find('video').text
+            if gamelist.game_tag_text_exist(xmlg, 'genre'):
+                self.genre = xmlg.find('genre').text
 
-            if game.find('releasedate') is not None:
-                self.releasedate = game.find('releasedate').text
+            if gamelist.game_tag_text_exist(xmlg, 'players'):
+                self.players = xmlg.find('players').text
 
-            if game.find('developer') is not None:
-                self.developer = game.find('developer').text
+            if gamelist.game_tag_text_exist(xmlg, 'core'):
+                self.core = xmlg.find('core').text
 
-            if game.find('publisher') is not None:
-                self.publisher = game.find('publisher').text
+            if gamelist.game_tag_text_exist(xmlg, 'emulator'):
+                self.emulator = xmlg.find('emulator').text
 
-            if game.find('genre') is not None:
-                self.genre = game.find('genre').text
+            if gamelist.game_tag_text_exist(xmlg, 'rating'):
+                self.rating = xmlg.find('rating').text
 
-            if game.find('players') is not None:
-                self.players = game.find('players').text
+            if gamelist.game_tag_text_exist(xmlg, 'playcount'):
+                self.playcount = xmlg.find('playcount').text
 
-            if game.find('core') is not None:
-                self.core = game.find('core').text
+            if gamelist.game_tag_text_exist(xmlg, 'lastplayed'):
+                self.lastplayed = xmlg.find('lastplayed').text
 
-            if game.find('emulator') is not None:
-                self.emulator = game.find('emulator').text
+            if gamelist.game_tag_text_exist(xmlg, 'md5'):
+                self.md5 = xmlg.find('md5').text
 
-            if game.find('rating') is not None:
-                self.rating = game.find('rating').text
-
-            if game.find('playcount') is not None:
-                self.playcount = game.find('playcount').text
-
-            if game.find('lastplayed') is not None:
-                self.lastplayed = game.find('lastplayed').text
-
-            if game.find('md5') is not None:
-                self.md5 = game.find('md5').text
+    def load_from_gamelist(self, gamelist, systempath):
+        self.load_media_from_gamelist(gamelist, systempath)
+        self.load_info_from_gamelist(gamelist)
 
 
 class System:
     """Base class for System in collection."""
 
-    def __init__(self, name, boxarts, images, marquees, thumbnails, videos):
-        self.name = name
+    def __init__(self, parent_dir, dir_name,
+                 boxarts=False, images=False, marquees=False,
+                 thumbnails=False, videos=False):
+        self.parent_dir = parent_dir
+        self.dir_name = dir_name
+        self.gamelist = None
         self.game = list()
-        self.boxarts = boxarts
-        self.images = images
-        self.marquees = marquees
-        self.thumbnails = thumbnails
-        self.videos = videos
+        self.box_dir = boxarts
+        self.img_dir = images
+        self.marq_dir = marquees
+        self.thumb_dir = thumbnails
+        self.vid_dir = videos
 
-    def save_to_dir(self, path):
-        pass
+    def get_path(self):
+        fh = FileHandler()
+        return fh.path([self.parent_dir, self.dir_name])
 
-    def read_from_dir(self, path):
-        pass
+    def get_gamelist_path(self):
+        fh = FileHandler()
+        return fh.path([self.get_path(), 'gamelist.xml'])
 
+    def get_file_list(self):
+        fh = FileHandler()
+        files = fh.get_files(self.get_path())
+        files.sort()
+        return files
 
-class GameListHandler:
-    """Base class for GameListXml operations."""
+    def load_gamelist(self):
+        self.gamelist = GameListHandler()
+        self.gamelist.set_tree(self.get_gamelist_path())
 
-    def __init__(self):
-        self.path = None
-        self.tree = None
-
-    def set_tree(self, path):
-        tree = None
-        if not os.path.isfile(path):
-            s = '<?xml version="1.0" encoding="UTF-8"?><gameList></gameList>'
-            f = open(path, "w")
-            f.write(s)
-            f.close()
-        tree = ET.parse(path)
-        self.tree = tree
-        self.path = path
-
-    def get_game_by_tag_text(self, tag, text):
-        if self.tree is not None:
-            for game in self.tree.getroot():
-                if game.find(tag) is not None:
-                    if game.find(tag).text is not None:
-                        tag_text = game.find(tag).text
-                        tag_text = tag_text.replace('./', '')
-                        text = tag_text.replace('./', '')
-                        if tag == 'path':
-                            tag_text = os.path.splitext(tag_text)[0]
-                            text = os.path.splitext(text)[0]
-                        if tag_text == text:
-                            return game
-        return False
-
-    def get_game_by_filename(self, filename):
-        return self.get_game_by_tag_text('path', filename)
+    def load_games(self):
+        for file in self.get_file_list():
+            game = Game(file)
+            game.load_from_gamelist(self.gamelist, self.get_path())
+            self.game.append(game)
 
 
-class Utils():
-    """docstring for Utils."""
+class Collection:
+    """Base class for System Collection."""
 
     def __init__(self):
-        super(Utils, self).__init__()
-        self.reset = '\033[0m'
-        self.bold = '\033[01m'
-        self.disable = '\033[02m'
-        self.underline = '\033[04m'
-        self.reverse = '\033[07m'
-        self.strikethrough = '\033[09m'
-        self.invisible = '\033[08m'
-        self.black = '\033[30m'
-        self.red = '\033[31m'
-        self.green = '\033[32m'
-        self.orange = '\033[33m'
-        self.blue = '\033[34m'
-        self.purple = '\033[35m'
-        self.cyan = '\033[36m'
-        self.darkgrey = '\033[90m'
-        self.yellow = '\033[93m'
-        self.pink = '\033[95m'
-
-        self.lightcyan = '\033[96m'
-        self.lightgrey = '\033[37m'
-        self.lightred = '\033[91m'
-        self.lightgreen = '\033[92m'
-        self.lightblue = '\033[94m'
-
-        self.bgblack = '\033[40m'
-        self.bgred = '\033[41m'
-        self.bggreen = '\033[42m'
-        self.bgorange = '\033[43m'
-        self.bgblue = '\033[44m'
-        self.bgpurple = '\033[45m'
-        self.bgcyan = '\033[46m'
-        self.bglightgrey = '\033[47m'
-
-    def str_red(self, text):
-        """docstring for Utils.print"""
-
-        return self.red + text + self.reset
-
-    def str_blue(self, text):
-        """docstring for Utils.print"""
-
-        return self.blue + text + self.reset
-
-    def str_green(self, text):
-        """docstring for Utils.print"""
-
-        return self.green + text + self.reset
+        pass